tg_call_bot/requirements.txt
# Telegram Bot - Requirements
aiogram>=3.0.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
openai>=1.0.0

# Notion API
notion-client>=2.0.0

# Optional dependencies
# redis>=4.0.0  # For Redis FSM storage
# sqlalchemy>=2.0.0  # For database operations
# alembic>=1.0.0  # For database migrations 